introduce TimeEntryControllerHelper
