+++ src/main/java/de/focusshift/zeiterfassung/timeentry/TimeEntryController.java
@@ -9,7 +9,6 @@ import de.focusshift.zeiterfassung.user.UserSettingsProvider;
 import de.focusshift.zeiterfassung.user.YearFormat;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
-import org.slf4j.Logger;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.core.oidc.user.OidcUser;
 import org.springframework.stereotype.Controller;
@@ -28,37 +27,32 @@ import java.time.Clock;
 import java.time.DayOfWeek;
 import java.time.Duration;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.time.Year;
-import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Locale;
 import java.util.Optional;
 
-import static java.lang.invoke.MethodHandles.lookup;
 import static java.time.Month.DECEMBER;
 import static java.time.temporal.IsoFields.WEEK_OF_WEEK_BASED_YEAR;
 import static java.time.temporal.TemporalAdjusters.previousOrSame;
-import static org.slf4j.LoggerFactory.getLogger;
 import static org.springframework.util.StringUtils.hasText;
 
 @Controller
 class TimeEntryController implements HasTimeClock, HasLaunchpad {
 
-    private static final Logger LOG = getLogger(lookup().lookupClass());
-
     private final TimeEntryService timeEntryService;
     private final UserSettingsProvider userSettingsProvider;
     private final DateFormatter dateFormatter;
+    private final TimeEntryControllerHelper timeEntryControllerHelper;
     private final Clock clock;
 
     public TimeEntryController(TimeEntryService timeEntryService, UserSettingsProvider userSettingsProvider,
-                               DateFormatter dateFormatter, Clock clock) {
+                               DateFormatter dateFormatter, TimeEntryControllerHelper timeEntryControllerHelper, Clock clock) {
         this.timeEntryService = timeEntryService;
         this.userSettingsProvider = userSettingsProvider;
         this.dateFormatter = dateFormatter;
+        this.timeEntryControllerHelper = timeEntryControllerHelper;
         this.clock = clock;
     }
 
@@ -91,7 +85,7 @@ class TimeEntryController implements HasTimeClock, HasLaunchpad {
     }
 
     private String prepareTimeEntriesView(int year, int weekOfYear, Model model, OidcUser principal, Locale locale) {
-        addTimeEntryToModel(model, new TimeEntryDTO());
+        timeEntryControllerHelper.addToModel(model, new TimeEntryDTO());
         addTimeEntriesToModel(year, weekOfYear, model, principal, locale);
 
         return "timeentries/index";
@@ -117,13 +111,14 @@ class TimeEntryController implements HasTimeClock, HasLaunchpad {
             throw new InvalidTimeEntryException("invalid time entry. date must be set.");
         }
 
-        final String viewName = saveOrUpdate(timeEntryDTO, bindingResult, model, principal, request);
+        timeEntryControllerHelper.saveOrUpdate(timeEntryDTO, bindingResult, model, principal);
 
         if (bindingResult.hasErrors()) {
             addTimeEntriesToModel(year, weekOfYear, model, principal, locale);
+            return "timeentries/index";
         }
 
-        return viewName;
+        return "redirect:" + request.getHeader("referer");
     }
 
     @PostMapping(value = "/timeentries/{id}")
@@ -148,7 +143,14 @@ class TimeEntryController implements HasTimeClock, HasLaunchpad {
             throw new InvalidTimeEntryException("invalid time entry. date must be set.");
         }
 
-        final String viewName = saveOrUpdate(timeEntryDTO, bindingResult, model, principal, request);
+        timeEntryControllerHelper.saveOrUpdate(timeEntryDTO, bindingResult, model, principal);
+
+        final String viewName;
+        if (bindingResult.hasErrors()) {
+            viewName = "timeentries/index";
+        } else {
+            viewName = "redirect:" + request.getHeader("referer");
+        }
 
         final UserId userId = new UserId(principal.getUserInfo().getSubject());
 
@@ -171,7 +173,7 @@ class TimeEntryController implements HasTimeClock, HasLaunchpad {
 
                 model.addAttribute("turboEditedWeek", toTimeEntryWeekDto(timeEntryWeek, locale));
                 model.addAttribute("turboEditedDay", toTimeEntryDayDto(timeEntryDay, locale));
-                model.addAttribute("turboEditedTimeEntry", toTimeEntryDto(editedTimeEntry));
+                model.addAttribute("turboEditedTimeEntry", timeEntryControllerHelper.toTimeEntryDto(editedTimeEntry));
             }
             return new ModelAndView("timeentries/index::#frame-time-entry");
         } else {
@@ -213,105 +215,9 @@ class TimeEntryController implements HasTimeClock, HasLaunchpad {
 
         model.addAttribute("turboEditedWeek", toTimeEntryWeekDto(timeEntryWeek, locale));
         model.addAttribute("turboEditedDay", timeEntryDay.map(entry -> toTimeEntryDayDto(entry, locale)).orElse(null));
-        model.addAttribute("turboDeletedTimeEntry", toTimeEntryDto(timeEntry));
+        model.addAttribute("turboDeletedTimeEntry", timeEntryControllerHelper.toTimeEntryDto(timeEntry));
 
         return "timeentries/index::#" + turboFrame;
-    }
-
-    private String saveOrUpdate(TimeEntryDTO dto, BindingResult bindingResult, Model model, OidcUser principal, HttpServletRequest request) {
-
-        if (bindingResult.hasErrors()) {
-            model.addAttribute("timeEntryErrorId", dto.getId());
-
-            final boolean hasErrorStart = bindingResult.hasFieldErrors("start");
-            final boolean hasErrorEnd = bindingResult.hasFieldErrors("end");
-            final boolean hasErrorDuration = bindingResult.hasFieldErrors("value");
-
-            if (hasErrorStart && hasErrorEnd && !hasErrorDuration) {
-                bindingResult.reject("time-entry.validation.startOrEnd.required");
-            } else if (hasErrorStart && !hasErrorEnd && hasErrorDuration) {
-                bindingResult.reject("time-entry.validation.startOrDuration.required");
-            } else if (!hasErrorStart && hasErrorEnd && hasErrorDuration) {
-                bindingResult.reject("time-entry.validation.endOrDuration.required");
-            }
-
-            return "timeentries/index";
-        }
-
-        final UserId userId = new UserId(principal.getUserInfo().getSubject());
-        final ZoneId zoneId = userSettingsProvider.zoneId();
-
-        if (dto.getId() == null) {
-            createTimeEntry(dto, userId, zoneId);
-        } else {
-            try {
-                updateTimeEntry(dto, zoneId);
-            } catch (TimeEntryUpdateNotPlausibleException e) {
-                LOG.debug("could not update time-entry", e);
-
-                bindingResult.reject("time-entry.validation.plausible");
-                bindingResult.rejectValue("start", "");
-                bindingResult.rejectValue("end", "");
-                bindingResult.rejectValue("duration", "");
-
-                model.addAttribute("timeEntryErrorId", dto.getId());
-
-                return "timeentries/index";
-            }
-        }
-
-        final String referer = request.getHeader("referer");
-        return "redirect:" + referer;
-    }
-
-    private void createTimeEntry(TimeEntryDTO dto, UserId userId, ZoneId zoneId) {
-
-        final ZonedDateTime start;
-        final ZonedDateTime end;
-
-        final Duration duration = toDuration(dto.getDuration());
-
-        if (duration.equals(Duration.ZERO)) {
-            // start and end should be given
-            start = ZonedDateTime.of(LocalDateTime.of(dto.getDate(), dto.getStart()), zoneId);
-            end = getEndDate(dto, zoneId);
-        } else if (dto.getStart() == null) {
-            // end and value should be given
-            end = ZonedDateTime.of(LocalDateTime.of(dto.getDate(), dto.getEnd()), zoneId);
-            start = end.minusMinutes(duration.toMinutes());
-        } else {
-            // start and value should be given
-            start = ZonedDateTime.of(LocalDateTime.of(dto.getDate(), dto.getStart()), zoneId);
-            end = start.plusMinutes(duration.toMinutes());
-        }
-
-        timeEntryService.createTimeEntry(userId, dto.getComment(), start, end, dto.isBreak());
-    }
-
-    private void updateTimeEntry(TimeEntryDTO dto, ZoneId zoneId) throws TimeEntryUpdateNotPlausibleException {
-
-        final Duration duration = toDuration(dto.getDuration());
-        final ZonedDateTime start = dto.getStart() == null ? null : ZonedDateTime.of(LocalDateTime.of(dto.getDate(), dto.getStart()), zoneId);
-        final ZonedDateTime end = getEndDate(dto, zoneId);
-
-        timeEntryService.updateTimeEntry(new TimeEntryId(dto.getId()), dto.getComment(), start, end, duration, dto.isBreak());
-    }
-
-    private ZonedDateTime getEndDate(TimeEntryDTO dto, ZoneId zoneId) {
-        if (dto.getEnd() == null) {
-            return null;
-        } else if (dto.getStart() == null) {
-            return ZonedDateTime.of(LocalDateTime.of(dto.getDate(), dto.getEnd()), zoneId);
-        } else if (dto.getEnd().isBefore(dto.getStart())) {
-            // end is on next day
-            return ZonedDateTime.of(LocalDateTime.of(dto.getDate().plusDays(1), dto.getEnd()), zoneId);
-        } else {
-            return ZonedDateTime.of(LocalDateTime.of(dto.getDate(), dto.getEnd()), zoneId);
-        }
-    }
-
-    private void addTimeEntryToModel(Model model, TimeEntryDTO timeEntryDTO) {
-        model.addAttribute("timeEntry", timeEntryDTO);
     }
 
     private void addTimeEntriesToModel(int year, int weekOfYear, Model model, OidcUser principal, Locale locale) {
@@ -386,7 +292,7 @@ class TimeEntryController implements HasTimeClock, HasLaunchpad {
         final String workedHoursShould = durationToTimeString(timeEntryDay.shouldWorkingHours().durationInMinutes());
         final Duration hoursDelta = timeEntryDay.overtime();
         final double ratio = timeEntryDay.workedHoursRatio().multiply(BigDecimal.valueOf(100), new MathContext(2)).doubleValue();
-        final List<TimeEntryDTO> dayTimeEntryDtos = timeEntryDay.timeEntries().stream().map(this::toTimeEntryDto).toList();
+        final List<TimeEntryDTO> dayTimeEntryDtos = timeEntryDay.timeEntries().stream().map(timeEntryControllerHelper::toTimeEntryDto).toList();
         final List<AbsenceEntryDto> absenceEntryDtos = timeEntryDay.absences().stream()
             .map(absence -> new AbsenceEntryDto(timeEntryDay.date(), absence.label(locale), absence.color()))
             .toList();
@@ -402,44 +308,6 @@ class TimeEntryController implements HasTimeClock, HasLaunchpad {
             .timeEntries(dayTimeEntryDtos)
             .absenceEntries(absenceEntryDtos)
             .build();
-    }
-
-    private TimeEntryDTO toTimeEntryDto(TimeEntry timeEntry) {
-
-        final ZonedDateTime start = timeEntry.start();
-        final ZonedDateTime end = timeEntry.end();
-
-        final LocalDate date = start.toLocalDate();
-        final LocalTime startTime = start.toLocalTime();
-        final LocalTime endTime = end.toLocalTime();
-
-        final Duration duration = timeEntry.durationInMinutes();
-        final String durationString = toTimeEntryDTODurationString(duration);
-
-        return TimeEntryDTO.builder()
-            .id(timeEntry.id().value())
-            .date(date)
-            .start(startTime)
-            .end(endTime)
-            .duration(durationString)
-            .comment(timeEntry.comment())
-            .isBreak(timeEntry.isBreak())
-            .build();
-    }
-
-    private static Duration toDuration(String timeEntryDTODurationString) {
-        if (hasText(timeEntryDTODurationString)) {
-            final String[] split = timeEntryDTODurationString.split(":");
-            return Duration.ofHours(Integer.parseInt(split[0])).plusMinutes(Integer.parseInt(split[1]));
-        }
-        return Duration.ZERO;
-    }
-
-    private static String toTimeEntryDTODurationString(Duration duration) {
-        if (duration == null) {
-            return "00:00";
-        }
-        return String.format("%02d:%02d", duration.toHours(), duration.toMinutes() % 60);
     }
 
     private int lastWeekOfYear(int year) {
